Refactor instance functions for readability
